+++ VulkanEngine/VEEngine.h
@@ -49,7 +49,7 @@ namespace ve {
 
 		double m_dt = 0.0;								///<Delta time since the last loop
 		double m_time = 0.0;							///<Absolute game time since start of the render loop
-		uint64_t m_loopCount = 0;						///<Counts up the render loop
+		uint32_t m_loopCount = 0;						///<Counts up the render loop
 		bool m_framebufferResized = false;				///<Flag indicating whether the window size has changed.
 		bool m_end_running = false;						///<Flag indicating that the engine should leave the render loop
 		bool m_debug = true;							///<Flag indicating whether debugging is enabled or not
@@ -87,7 +87,7 @@ namespace ve {
 		VEWindow       * getWindow();				//Return a pointer to the window instance
 		VESceneManager * getSceneManager();			//Return a  pointer to the scene manager instance
 		VERenderer     * getRenderer();				//Return a pointer to the renderer instance
-		uint64_t		 getLoopCount();			//Return the number of the current render loop
+		uint32_t		 getLoopCount();			//Return the number of the current render loop
 	};
